[Refactoring ⚙️] 불필요한 파일 제거및 파일 경로 오류 수정
