+++ README.md
@@ -3,5 +3,4 @@
 ## 테스트 계정
 
 id : gnsdh8616@konkuk.ac.kr
-
 password: 123123gg
