--- src/components/common/Avatar/Avatar.stories.tsx
@@ -1,15 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react"
-
-import Avatar from "@components/common/Avatar/Avatar"
-
-const meta: Meta<typeof Avatar> = {
-  component: Avatar,
-  title: "components/Avatar",
-  tags: ["autodocs"],
-  parameters: { layout: "centered" },
-}
-
-export default meta
-type Story = StoryObj<typeof Avatar>
-
-export const Primary: Story = {}
