--- src/components/common/Avatar/Avatar.tsx
@@ -1,10 +1,0 @@
-import Icon from "@components/common/Icon/Icon"
-
-const Avatar = () => (
-  <Icon
-    icon="Avatar"
-    size={130}
-  />
-)
-
-export default Avatar
