+++ src/components/common/Card/Card.stories.tsx
@@ -17,8 +17,8 @@ export const Primary: Story = {
     <Card
       src="https://fitness-mate.s3.ap-northeast-2.amazonaws.com/images/workout/2d844f66-72a2-4189-99f0-bdf3a38d2963.png"
       title="런지"
-      badges={["등", "가슴"]}>
-      {" "}
-    </Card>
+      badges={["등", "가슴"]}
+      style={{ width: "30rem" }}
+    />
   ),
 }
