+++ src/components/common/Title/Title.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react"
 
-import Title from "@components/Title/Title"
+import Title from "@components/common/Title/Title"
 
 const meta: Meta<typeof Title> = {
   title: "components/Title",
