+++ src/hooks/mutation/useLogout.ts
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom"
 
-import { useUserStore } from "@store/useUserStore"
+import { useUserStore } from "stores/useUserStore"
 
 import { useMutation } from "@tanstack/react-query"
