+++ src/layouts/Mainlayout.tsx
@@ -3,11 +3,11 @@ import { Outlet, useLocation } from "react-router-dom"
 import { ToastOptions } from "react-toastify"
 import "react-toastify/dist/ReactToastify.css"
 
-import { useModalStore } from "@store/useModalStore"
-import { useUserStore } from "@store/useUserStore"
+import Footer from "layouts/Footer/Footer"
+import Navbar from "layouts/Navbar/Navbar"
+import { useModalStore } from "stores/useModalStore"
+import { useUserStore } from "stores/useUserStore"
 
-import Footer from "@components/Footer/Footer"
-import Navbar from "@components/Navbar/Navbar"
 import { ScrollToTop } from "@components/ScrollToTop/ScrollToTop"
 import { StyledToast } from "@components/Toast/Toast"
 import AlertLoadingModal from "@components/common/Modal/components/Alert/AlertLoadingModal"
