+++ src/pages/Signup/Progressbar/StyledProgressBar.ts
