+++ src/stories/icon.mdx
@@ -1,7 +1,6 @@
 import { IconGallery, IconItem, Meta, Title } from "@storybook/blocks"
 
-import { ICON_MAP } from "../components/Icon/Icon.tsx"
-import Icon from "../components/Icon/Icon.tsx"
+import Icon, { ICON_MAP } from "../components/common/Icon/Icon"
 
 <Meta title="Getting Started/Icon" />
